% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootG3D.r
\name{bootG3D}
\alias{bootG3D}
\title{Bootstrap G3D fitting procedure to get parameter distributions and CI}
\usage{
bootG3D(data, N, a, skipfirst = 4)
}
\arguments{
\item{data}{dataframe}

\item{indices}{}
}
\description{
Bootstrap G3D fitting procedure to get parameter distributions and CI
}
\examples{
require(boot)
fcsData <- stack_fcs(path.dir)
results <- boot(data=fcsmed[c(-1:-20),], statistic=bootG3D, init=list(N=45,td=.1),R=10)
bootres <- bootG3D(fcsData, N=500, a=9)
#plot FCS & fit
rbind(apply(bootres,1,median),apply(bootres,1,quantile,c(.025,.975)))
hist(bootres[1,], breaks=20)
hist(bootres[2,], breaks=20)
plot_fcsfit(frm=dss[c(-1:-20),],fit=G3D,bootres=results)
}
\author{
JuG
}
