% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G3D_fit.r
\name{G3D_fit}
\alias{G3D_fit}
\title{Fit autocorrelation curve assuming free 3D diffusion in a 3‐dimensional Gaussian excitation volume}
\usage{
G3D_fit(frm, init)
}
\arguments{
\item{frm}{dataframe with Tp (lagtime) and G (autocorrelation) data}

\item{init}{initialization parameters list}
}
\value{
nls object
}
\description{
$G = \frac{1}{N}   \left( 1 + \frac{Tp}{td} \right)^{-1}  \left(1 + \frac{Tp}{td/9} \right)^{-1/2}$
}
\examples{
G3D <-G3D_fit(dss[c(-4:-1),],init=list(N=3,td=1))  #td en msec
print(summary(G3D) )
}
\author{
JuG
}
