% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G3Dt_fit.r
\name{G3Dt_fit}
\alias{G3Dt_fit}
\title{Fit autocorrelation curve assuming free 3D diffusion with triplet state in a 3‐dimensional Gaussian excitation volume}
\usage{
G3Dt_fit(frm, init, a = 9)
}
\arguments{
\item{frm}{dataframe with Tp (lagtime) and G (autocorrelation) data}

\item{init}{initialization parameters list}
}
\value{
nls object
}
\description{
$G = \frac{1}{N}   \left( 1 + \frac{Tp}{td} \right)^{-1}  \left(1 + \frac{Tp}{td/9} \right)^{-1/2}  \left(1 + \frac{ft}{1 - ft} exp(-Tp/tt) \right)$
}
\examples{
G3Dt <-G3Dt_fit(dss[c(-1:-4),],init=list(N=50,td=0.06,ft=.75,tt=.01))
print(summary(G3Dt) )
#check fit diagnostic
if(!require(nlstools)){install.packages('nlstools')}
library(nlstools)
plot(nlsResiduals (G3Dt))
}
\author{
JuG
}
