% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fcs.r
\name{plot_fcs}
\alias{plot_fcs}
\title{Plot FCS curve}
\usage{
plot_fcs(frm, addFit = F, model, allCurves = F, ...)
}
\arguments{
\item{frm}{dataframe with Tp (Tau in ms) aud G (G(tau))}
}
\description{
Plot FCS curve
}
\examples{
plot_fcs(fcsmed)
plot_fcs(fcsmed,xlim=c(2e-3, 100), ylim = c(0,.04))
fcsData <- stack_fcs(path.dir)
fcsmed <- data.frame(Tp = fcsData$Time, G = apply(fcsData[,-1],1,median))
fcsG <- cbind(fcsmed,fcsData[,-1])
G3Dt <-G3D_fit(fcsmed[-10:-1,],init=list(N=3,td=1), a = 9.4)
plot_fcs(fcsmed,xlim=c(1e-3, 100), type='l',ylim = c(0,.04),addFit=T, model=G3D)
plot_fcs(fcsG,xlim=c(1e-3, 100), type='l',ylim = c(0,.04),addFit=T, model=G3D, allCurves=T)
}
\author{
JuG
}
