% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_CR.r
\name{read_CR}
\alias{read_CR}
\title{#' Read Count Rate bloc in  ALV .asc files}
\usage{
read_CR(path)
}
\arguments{
\item{path}{path to the .asc file}
}
\value{
data.frame
}
\description{
Read count rate block in AVL .asc files.
}
\details{

}
\examples{
flist <- list.files(path=paste(getwd(),"/",path.dir,sep=''), pattern="[.]ASC$",full.names = TRUE )
CR <- read_CR(path=flist[1])
plot(CR, type='l', col='blue',ylim=c(0,1.1*max(CR[,2])))
}
\author{
JuG
}
