% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fcs.r
\name{read_fcs}
\alias{read_fcs}
\title{Read correlation block in ALV .asc files}
\usage{
read_fcs(path, nrows = 181)
}
\arguments{
\item{path}{path to the .asc file}

\item{nrows}{number of lines of the correlation data block}
}
\value{
data.frame (Tp == lagtime,  G == autocorrelation curve)
}
\description{
Read correlation block in ALV .asc files
}
\examples{

flist <- list.files(path=paste(getwd(),"/",path.dir,sep=''), pattern="[.]ASC$",full.names = TRUE )
dss <- read_fcs(path=flist[1])
}
\author{
JuG
}
