% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_cr.R
\name{stack_cr}
\alias{stack_cr}
\title{Read count rate block in ALV for all .asc files in a directory}
\usage{
stack_cr(path.dir, ...)
}
\arguments{
\item{path.dir}{path to directory containing the .asc files}
}
\description{
Read count rate block in ALV for all .asc files in a directory
}
\examples{
path.dir <- paste(getwd(),"/data",sep='')
fcscr <- stack_cr(path.dir)
#median curve
crmed <- data.frame(Tp = fcscr$Time, G = apply(fcscr[,-1],1,median))
plot(fcscr[,1:2],type='l',ylim=c(0,150))
for (i in 2:(dim(fcscr)[2])){
    lines(fcscr[,c(1,i)])
}

}
\author{
JuG
}
