% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_fcs.r
\name{stack_fcs}
\alias{stack_fcs}
\title{Read correlation block in ALV for all .asc files in a directory}
\usage{
stack_fcs(path.dir)
}
\arguments{
\item{path.dir}{path to directory containing the .asc files}
}
\description{
Read correlation block in ALV for all .asc files in a directory
}
\examples{
path.dir <- paste(getwd(),"/data",sep='')
fcsData <- stack_fcs(path.dir)
#median curve
fcsmed <- data.frame(Tp = fcsData$Time, G = apply(fcsData[,-1],1,median))
G3D <- G3D_fit(fcsmed[c(-1:-4),],init=list(N=30,td=0.1),a=9)
plot_fcs(fcsmed,addFit = T,model = G3D,ylim=c(0,.04))
}
\author{
JuG
}
