% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G3D_fit.r
\name{G3D_fit}
\alias{G3D_fit}
\title{Fit autocorrelation curve assuming free 3D diffusion in a 3‐dimensional Gaussian excitation volume}
\usage{
G3D_fit(frm, init, a, ...)
}
\arguments{
\item{frm}{dataframe with Tp (lagtime) and G (autocorrelation) data}

\item{init}{initialization parameters list}
}
\value{
nls object
}
\description{
$G = \frac{1}{N}   \left( 1 + \frac{Tp}{td} \right)^{-1}  \left(1 + \frac{Tp}{td/9} \right)^{-1/2}$
}
\examples{
#'a' floating
G3D_fit(fcsmed[-10:-1,],init=list(N=3,td=1, a=9), algorithm = "port", lower=list(N=0, td=0, a=0), upper=list(N=1000, td=10, a=10))
# a fixed
G3Dt <-G3D_fit(fcsmed[-10:-1,],init=list(N=3,td=1), a = 9.4)
print(summary(G3D) )
cbind(coef(G3D),confint(G3D))
#check fit diagnostic
if(!require(nlstools)){install.packages('nlstools')}
library(nlstools)
plot(nlsResiduals (G3D))
}
\author{
JuG
}
