% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fcs.r
\name{read_fcs}
\alias{read_fcs}
\title{Read ALV .asc files}
\usage{
read_fcs(path, nrows = 334)
}
\arguments{
\item{path}{path to the .asc file}

\item{nrows}{number of lines of the correlation data block}
}
\value{
data.frame (Tp == lagtime,  G == autocorrelation curve)
}
\description{
Read ALV .asc files
}
\examples{

}
\author{
JuG
}
